# Improve perf of Utf8Parser.TryParse for int64 and uint64