Somewhat inspired by investigating https://github.com/dotnet/runtime/issues/52314 and seeing a little bit of low-hanging fruit that we can address.

Perf results (basically from running [this benchmark](https://github.com/dotnet/performance/blob/716f16c4ce263d03e6fcf1ea8e60bcf5fcbee600/src/benchmarks/micro/libraries/System.Buffers/Utf8ParserTests.cs#L22) with some additional inputs):

|        Method | Toolchain |                value |     Mean |     Error |   StdDev | Ratio |
|-------------- |---------- |--------------------- |---------:|----------:|---------:|------:|
| TryParseInt64 |  baseline |                   -1 |  3.73 ns |  0.022 ns | 0.018 ns |  1.00 |
| TryParseInt64 |   compare |                   -1 |  3.25 ns |  0.025 ns | 0.021 ns |  0.87 |
|               |           |                      |          |           |          |       |
| TryParseInt64 |  baseline |               -12345 |  5.93 ns |  1.680 ns | 0.092 ns |  1.00 |
| TryParseInt64 |   compare |               -12345 |  5.85 ns |  1.097 ns | 0.060 ns |  0.99 |
|               |           |                      |          |           |          |       |
| TryParseInt64 |  baseline |   -12345678901234567 | 15.75 ns | 43.214 ns | 2.368 ns |  1.00 |
| TryParseInt64 |   compare |   -12345678901234567 | 13.24 ns |  0.371 ns | 0.020 ns |  0.85 |
|               |           |                      |          |           |          |       |
| TryParseInt64 |  baseline | -9223372036854775808 | 19.35 ns |  0.326 ns | 0.305 ns |  1.00 |
| TryParseInt64 |   compare | -9223372036854775808 | 17.12 ns |  0.085 ns | 0.076 ns |  0.88 |
|               |           |                      |          |           |          |       |
| TryParseInt64 |  baseline |                    0 |  3.17 ns |  0.097 ns | 0.005 ns |  1.00 |
| TryParseInt64 |   compare |                    0 |  3.21 ns |  0.602 ns | 0.033 ns |  1.01 |
|               |           |                      |          |           |          |       |
| TryParseInt64 |  baseline |                    1 |  3.27 ns |  0.060 ns | 0.003 ns |  1.00 |
| TryParseInt64 |   compare |                    1 |  3.21 ns |  0.322 ns | 0.017 ns |  0.98 |
|               |           |                      |          |           |          |       |
| TryParseInt64 |  baseline |                12345 |  5.76 ns |  0.738 ns | 0.040 ns |  1.00 |
| TryParseInt64 |   compare |                12345 |  5.07 ns |  0.929 ns | 0.050 ns |  0.88 |
|               |           |                      |          |           |          |       |
| TryParseInt64 |  baseline |    12345678901234567 | 13.90 ns |  0.247 ns | 0.013 ns |  1.00 |
| TryParseInt64 |   compare |    12345678901234567 | 12.85 ns |  1.504 ns | 0.082 ns |  0.92 |
|               |           |                      |          |           |          |       |
| TryParseInt64 |  baseline |  9223372036854775807 | 17.72 ns |  0.102 ns | 0.095 ns |  1.00 |
| TryParseInt64 |   compare |  9223372036854775807 | 16.11 ns |  0.063 ns | 0.059 ns |  0.91 |

Some of the benchmarks were going a bit screwy on my machine, so I basically ran them one at a time and concatenated all of the results into a single table. Not sure what was up with my environment.

The tl;dr is that this shouldn't have any worse perf than the existing code, and the total codegen ends up being __448 bytes smaller__ (on x64) compared to baseline across the two methods.