@GrabYourPitchforks 2021-07-12
I investigated the framework's other parsing routines, but they have a bit more complexity, needing to handle whitespace and trailing nulls. Opened https://github.com/dotnet/runtime/issues/55541 so we don't lose track of this.